cml {multcomp}	R Documentation

Chronic Myelogenous Leukemia survival data.

Description

Survival in a randomised trial comparing three treatments for Chronic Myelogeneous Leukemia (simulated data).

Usage

data("cml")

Format

A data frame with 507 observations on the following 7 variables.

center

a factor with 54 levels indicating the study center.

treatment

a factor with levels trt1, trt2, trt3 indicating the treatment group.

sex

sex (0 = female, 1 = male)

age

age in years

riskgroup

risk group (0 = low, 1 = medium, 2 = high)

status

censoring status (FALSE = censored, TRUE = dead)

time

survival or censoring time in days.

Details

The data are simulated according to structure of the data by the German CML Study Group used in Hehlmann, Heimpel, Hasford, Kolb, Pralle, Hossfeld, Queisser, Loffler, Hochhaus, and Heinze (1994).

References

Hehlmann R, Heimpel H, Hasford J, Kolb H., Pralle H, Hossfeld D., Queisser W, Loffler H, Hochhaus A, Heinze B (1994). “Randomized Comparison of Interferon-alpha with Busulfan and Hydroxyurea in Chronic Myelogenous Leukemia. The German CML Study Group.” Blood, 84(12), 4064–4077. doi: 10.1182/blood.V84.12.4064.bloodjournal84124064.

Examples

if (require("coxme")) {
    data("cml")
    ### one-sided simultaneous confidence intervals for many-to-one 
    ### comparisons of treatment effects concerning time of survival 
    ### modeled by a frailty Cox model with adjustment for further 
    ### covariates and center-specific random effect.
    cml_coxme <- coxme(Surv(time, status) ~ treatment + sex + age + riskgroup + (1|center), 
                       data = cml)
    glht_coxme <- glht(model = cml_coxme, linfct = mcp(treatment = "Dunnett"), 
                       alternative = "greater")
    ci_coxme <- confint(glht_coxme)
    exp(ci_coxme$confint)[1:2,]
}

[Package multcomp version 1.4-30 Index]