detergent {multcomp}	R Documentation

Detergent Durability Data Set

Description

Detergent durability in an incomplete two-way design.

Usage

data("detergent")

Format

This data frame contains the following variables

detergent

detergent, a factor at levels A, B, C, D, and E.

block

block, a factor at levels B_1, ..., B_10.

plates

response variable: number of plates washed before the foam disappears.

Details

Plates were washed with five detergent varieties, in ten blocks. A complete design would have 50 combinations, here only three detergent varieties in each block were applied in a balanced incomplete block design. Note that there are six observations taken at each detergent level.

Source

H. Scheffe (1959). The Analysis of Variance. New York: John Wiley & Sons, page 189.

P. H. Westfall, R. D. Tobias, D. Rom, R. D. Wolfinger, Y. Hochberg (1999). Multiple Comparisons and Multiple Tests Using the SAS System. Cary, NC: SAS Institute Inc., page 189.

Examples

  ### set up two-way ANOVA without interactions
  amod <- aov(plates ~ block + detergent, data = detergent)

  ### set up all-pair comparisons
  dht <- glht(amod, linfct = mcp(detergent = "Tukey"))

  ### see Westfall et al. (1999, p. 190)
  confint(dht)

  ### see Westfall et al. (1999, p. 192)
  summary(dht, test = univariate())
  ## Not run: 
  summary(dht, test = adjusted("Shaffer"))
  summary(dht, test = adjusted("Westfall"))
  
## End(Not run)

[Package multcomp version 1.4-19 Index]