fattyacid {multcomp}	R Documentation

Fatty Acid Content of Bacillus simplex.

Description

Fatty acid content of different putative ecotypes of Bacillus simplex.

Usage

data("fattyacid")

Format

A data frame with 93 observations on the following 2 variables.

PE

a factor with levels PE3, PE4, PE5, PE6, PE7, PE9 indicating the putative ecotype (PE).

FA

a numeric vector indicating the content of fatty acid (FA).

Details

The data give the fatty acid content for different putative ecotypes of Bacillus simplex. Variances of the values of fatty acid are heterogeneous among the putative ecotypes.

Source

J. Sikorski, E. Brambilla, R. M. Kroppenstedt, B. J. Tindal (2008), The temperature adaptive fatty acid content in Bacillus simplex strains from ”Evolution Canyon“, Israel. Microbiology 154, 2416-2426.

Examples

if (require("sandwich")) {
    data("fattyacid")
    ### all-pairwise comparisons of the means of fatty acid content 
    ### FA between different putative ecotypes PE accounting for 
    ### heteroscedasticity by using a heteroscedastic consistent 
    ### covariance estimation
    amod <- aov(FA ~ PE, data = fattyacid)
    amod_glht <- glht(amod, mcp(PE = "Tukey"), vcov = vcovHC)
    summary(amod_glht)

    ### simultaneous confidence intervals for the differences of 
    ### means of fatty acid content between the putative ecotypes
    confint(amod_glht)
}

[Package multcomp version 1.4-19 Index]