litter {multcomp}	R Documentation

Litter Weights Data Set

Description

Dose response of litter weights in rats.

Usage

data("litter")

Format

This data frame contains the following variables

dose

dosages at four levels: 0, 5, 50, 500.

gesttime

gestation time as covariate.

number

number of animals in litter as covariate.

weight

response variable: average post-birth weights in the entire litter.

Details

Pregnant mice were divided into four groups and the compound in four different doses was administered during pregnancy. Their litters were evaluated for birth weights, see Westfall, Tobias, Rom, Wolfinger, and Hochberg (1999), page 109, and Westfall (1997).

References

Westfall P. H. (1997). “Multiple Testing of General Contrasts using Logical Constraints and Correlations.” Journal of the American Statistical Association, 92(437), 299–306. doi: 10.1080/01621459.1997.10473627.

Westfall P. H., Tobias R. D., Rom D., Wolfinger R. D., Hochberg Y. (1999). Multiple Comparisons and Multiple Tests Using the SAS System. SAS Institute Inc., Cary, NC.

Examples

  ### fit ANCOVA model to data
  amod <- aov(weight ~ dose + gesttime + number, data = litter)

  ### define matrix of linear hypotheses for `dose'
  doselev <- as.integer(levels(litter$dose))
  K <- rbind(contrMat(table(litter$dose), "Tukey"),
             otrend = c(-1.5, -0.5, 0.5, 1.5),
             atrend = doselev - mean(doselev),
             ltrend = log(1:4) - mean(log(1:4)))

  ### set up multiple comparison object
  Kht <- glht(amod, linfct = mcp(dose = K), alternative = "less")

  ### cf. Westfall (1997, Table 2)
  summary(Kht, test = univariate())
  summary(Kht, test = adjusted("bonferroni"))
  summary(Kht, test = adjusted("Shaffer"))
  summary(Kht, test = adjusted("Westfall"))
  summary(Kht, test = adjusted("single-step"))

[Package multcomp version 1.4-30 Index]