parm {multcomp}	R Documentation

Model Parameters

Description

Directly specify estimated model parameters and their covariance matrix.

Usage

parm(coef, vcov, df = 0)

Arguments

coef
estimated coefficients.

vcov
estimated covariance matrix of the coefficients.

df
an optional specification of the degrees of freedom to be used in subsequent computations.

Details

When only estimated model parameters and the corresponding covariance matrix is available for simultaneous inference using glht (for example, when only the results but not the original data are available or, even worse, when the model has been fitted outside R), function parm sets up an object glht is able to compute on (mainly by offering coef and vcov methods).

Note that the linear function in glht can't be specified via mcp since the model terms are missing.

Value

An object of class parm with elements

coef
model parameters

vcov
covariance matrix of model parameters

df
degrees of freedom

Examples

## example from
## Bretz, Hothorn, and Westfall (2002). 
## On multiple comparisons in R. R News, 2(3):14-17.

beta <- c(V1 = 14.8, V2 = 12.6667, V3 = 7.3333, V4 = 13.1333)
Sigma <- 6.7099 * (diag(1 / c(20, 3, 3, 15)))
confint(glht(model = parm(beta, Sigma, 37),
             linfct = c("V2 - V1 >= 0", 
                        "V3 - V1 >= 0", 
                        "V4 - V1 >= 0")), 
        level = 0.9)

[Package multcomp version 1.4-19 Index]