trees513 {multcomp}	R Documentation

Frankonian Tree Damage Data

Description

Damages on young trees caused by deer browsing.

Usage

data("trees513")

Format

A data frame with 2700 observations on the following 4 variables.

damage

a factor with levels yes and no indicating whether or not the trees has been damaged by game animals, mostly roe deer.

species

a factor with levels spruce, fir, pine, softwood (other), beech, oak, ash/maple/elm/lime, and hardwood (other).

lattice

a factor with levels 1, ..., 53, essentially a number indicating the position of the sampled area.

plot

a factor with levels x_1, ..., x_5 where x is the lattice. plot is nested within lattice and is a replication for each lattice point.

Details

In most parts of Germany, the natural or artificial regeneration of forests is difficult due to a high browsing intensity. Young trees suffer from browsing damage, mostly by roe and red deer. In order to estimate the browsing intensity for several tree species, the Bavarian State Ministry of Agriculture and Foresty conducts a survey every three years. Based on the estimated percentage of damaged trees, suggestions for the implementation or modification of deer management plans are made. The survey takes place in all 756 game management districts (‘Hegegemeinschaften’) in Bavaria. The data given here are from the game management district number 513 ‘Unterer Aischgrund’ (located in Frankonia between Erlangen and Höchstadt) in 2006. The data of 2700 trees include the species and a binary variable indicating whether or not the tree suffers from damage caused by deer browsing.

Data and analysis are presented by Hothorn, Bretz, and Westfall (2008).

References

Hothorn T., Bretz F., Westfall P. (2008). “Simultaneous Inference in General Parametric Models.” Biometrical Journal, 50(3), 346–363. doi: 10.1002/bimj.200810425.

Examples

  summary(trees513)

[Package multcomp version 1.4-30 Index]